feat: 토너먼트 결과 페이지 GET /{id} 연동
